5758 - Add meta/socket/event/links.ts
