+++ src/meta/socket/event/links.ts
@@ -1,0 +1,6 @@
+export enum LinksVerificationEvent {
+  queued = 'queued',
+  active = 'active',
+  completed = 'completed',
+  failed = 'failed',
+}
