+++ src/meta/socket/sockets/links.ts
@@ -2,13 +2,6 @@ import { CountryIso } from 'meta/area/countryIso'
 import { AssessmentName } from 'meta/assessment/assessment'
 import { CycleName } from 'meta/assessment/cycle'
 
-export enum LinksVerificationEvent {
-  queued = 'queued',
-  active = 'active',
-  completed = 'completed',
-  failed = 'failed',
-}
-
 type Props = {
   assessmentName: AssessmentName
   countryIso?: CountryIso
